refactor: move headers and upload params generators into Param namespace
