+++ lib/uploadcare/uploadcare.rb
@@ -10,7 +10,7 @@ require 'entity/entity'
 Gem.find_files('client/**/*.rb').each { |path| require path }
 Gem.find_files('concerns/**/*.rb').each { |path| require path }
 Gem.find_files('entity/**/*.rb').each { |path| require path }
-Gem.find_files('header/**/*.rb').each { |path| require path }
+Gem.find_files('param/**/*.rb').each { |path| require path }
 Gem.find_files('service/**/*.rb').each { |path| require path }
 require 'api/api'
