+++ spec/uploadcare/param/authentication_header_spec.rb
