+++ spec/uploadcare/param/secure_auth_header_spec.rb
