+++ spec/uploadcare/param/simple_auth_header_spec.rb
